Date: 4/08/2025 commit

+++ src/main/java/com/ecommerce/belgro/Model/Order.java
@@ -39,7 +39,7 @@ public class Order {
 
     private double totalMrpPrice;
 
-    private Integer totalSellingPrice;
+    private double totalSellingPrice;
 
     private Integer discount;
