+++ src/main/java/com/ecommerce/belgro/Model/PaymentOrder.java
@@ -19,7 +19,7 @@ public class PaymentOrder {
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long id;
 
-    private Long amount;
+    private double amount;
     private PaymentOrderStatus status = PaymentOrderStatus.PENDING;
     private PaymentMethod paymentMethod;
     private String paymentLinkId;
